+++ lib/browser/assertions/assert_elements.ts
@@ -1,4 +1,4 @@
-import {AssertionError} from '../../errors';
+import { AssertionError } from '../../errors';
 import { isNumber } from '../../utils/utils';
 import { WendigoSelector } from '../../types';