+++ lib/modules/local_storage/local_storage_not_assertions.ts
@@ -1,4 +1,4 @@
-import * as assertUtils from '../../utils/assert_utils';
+import { invertify } from '../../utils/assert_utils';
 import BrowserLocalStorageAssertions from './local_storage_assertions';
 import { arrayfy } from '../../utils/utils';
 
@@ -15,7 +15,7 @@ export default class BrowserLocalStorageNotAssertions {
             const itemText = keyList.length === 1 ? "item" : "items";
             msg = `Expected ${itemText} "${keyList.join(" ")}" not to exist in localStorage.`;
         }
-        return assertUtils.invertify(() => {
+        return invertify(() => {
             return this.localStorageAssertions.exist(keyList, "");
         }, "assert.localStorage.not.exist", msg);
     }
@@ -37,7 +37,7 @@ export default class BrowserLocalStorageNotAssertions {
             msg = `Expected ${itemText} "${keys.join(" ")}" not to have ${valuesText} "${expectedVals}" in localStorage.`;
         }
 
-        return assertUtils.invertify(() => {
+        return invertify(() => {
             return this.localStorageAssertions.value(keyVals, "");
         }, "assert.localStorage.not.value", msg);
     }
@@ -47,14 +47,14 @@ export default class BrowserLocalStorageNotAssertions {
             const itemText = expected === 1 ? "item" : "items";
             msg = `Expected localStorage not to have ${expected} ${itemText}.`;
         }
-        return assertUtils.invertify(() => {
+        return invertify(() => {
             return this.localStorageAssertions.length(expected, "");
         }, "assert.localStorage.not.length", msg);
     }
 
     public empty(msg?: string): Promise<void> {
         if (!msg) msg = `Expected localStorage not to be empty.`;
-        return assertUtils.invertify(() => {
+        return invertify(() => {
             return this.localStorageAssertions.empty("");
         }, "assert.localStorage.not.empty", msg);
     }